backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pytesseract==0.3.13
easyocr==1.7.1
pdf2image==1.17.0
qrcode[pil]==7.4.2
bcrypt
python-jose
google-auth
httpx>=0.27.0
beautifulsoup4>=4.12.0
playwright==1.58.0
